API/requirements.txt
# Web Framework
tornado>=6.4

# MongoDB
motor>=3.3.2
pymongo>=4.6.1

# JWT Authentication
PyJWT>=2.8.0
bcrypt>=4.1.2

# Image Processing
Pillow>=10.2.0
numpy>=1.26.0,<2.0.0
opencv-python-headless>=4.8.0
scipy>=1.11.0

# Deep Learning (Real-ESRGAN)
torch>=2.0.0
torchvision>=0.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
email-validator>=2.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
